Release : 2.3.2

+++ miscellaneoustool/data/src/main/kotlin/com/jinproject/data/datasource/cache/database/entity/Item.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 data class Item(
     @PrimaryKey val itemName: String,
     val itemType: String,
-    val itemPrice: Int
+    val itemPrice: Long,
 )
